refactor(compiler): reorganizing and renaming the executors
