+++ include/Ark/Compiler/Macros/Executors/Conditional.hpp
@@ -1,16 +1,16 @@
 /**
- * @file ListExecutor.hpp
+ * @file Conditional.hpp
  * @author Ray John Alovera (rakista112@gmail.com)
- * @brief Executor for List Macros
- * @version 0.3
+ * @brief Executor for Conditional Macros
+ * @version 0.4
  * @date 2021-05-04
  * 
  * @copyright Copyright (c) 2021
  * 
  */
 
-#ifndef MACROS_EXECUTORS_LISTEXECUTOR_HPP
-#define MACROS_EXECUTORS_LISTEXECUTOR_HPP
+#ifndef MACROS_EXECUTORS_CONDITIONAL_HPP
+#define MACROS_EXECUTORS_CONDITIONAL_HPP
 
 #include <Ark/Compiler/Macros/Executor.hpp>
 #include <Ark/Compiler/AST/Node.hpp>
@@ -18,13 +18,13 @@
 namespace Ark::internal
 {
     /**
-     * @brief Handles List macros
+     * @brief Handles Conditional macros
      * 
     */
-    class ListExecutor : public MacroExecutor
+    class ConditionalExecutor : public MacroExecutor
     {
     public:
-        ListExecutor(MacroProcessor* macroprocessor, unsigned debug = 0);
+        using MacroExecutor::MacroExecutor;
 
         bool applyMacro(Node& node) override;
         bool canHandle(Node& node) override;
