+++ include/Ark/Compiler/Macros/Executors/List.hpp
@@ -1,16 +1,16 @@
 /**
- * @file ConditionalExecutor.hpp
+ * @file List.hpp
  * @author Ray John Alovera (rakista112@gmail.com)
- * @brief Executor for Conditional Macros
- * @version 0.4
+ * @brief Executor for List Macros
+ * @version 0.3
  * @date 2021-05-04
  * 
  * @copyright Copyright (c) 2021
  * 
  */
 
-#ifndef MACROS_EXECUTORS_CONDITIONALEXECUTOR_HPP
-#define MACROS_EXECUTORS_CONDITIONALEXECUTOR_HPP
+#ifndef MACROS_EXECUTORS_LIST_HPP
+#define MACROS_EXECUTORS_LIST_HPP
 
 #include <Ark/Compiler/Macros/Executor.hpp>
 #include <Ark/Compiler/AST/Node.hpp>
@@ -18,13 +18,13 @@
 namespace Ark::internal
 {
     /**
-     * @brief Handles Conditional macros
+     * @brief Handles List macros
      * 
     */
-    class ConditionalExecutor : public MacroExecutor
+    class ListExecutor : public MacroExecutor
     {
     public:
-        ConditionalExecutor(MacroProcessor* macroprocessor, unsigned debug = 0);
+        using MacroExecutor::MacroExecutor;
 
         bool applyMacro(Node& node) override;
         bool canHandle(Node& node) override;
