+++ include/Ark/Compiler/Macros/Executors/Symbol.hpp
@@ -1,5 +1,5 @@
 /**
- * @file SymbolExecutor.hpp
+ * @file Symbol.hpp
  * @author Ray John Alovera (rakista112@gmail.com)
  * @brief Executor for Symbol Macros
  * @version 0.3
@@ -9,8 +9,8 @@
  * 
  */
 
-#ifndef MACROS_EXECUTORS_SYMBOLEXECUTOR_HPP
-#define MACROS_EXECUTORS_SYMBOLEXECUTOR_HPP
+#ifndef MACROS_EXECUTORS_SYMBOL_HPP
+#define MACROS_EXECUTORS_SYMBOL_HPP
 
 #include <Ark/Compiler/Macros/Executor.hpp>
 #include <Ark/Compiler/AST/Node.hpp>
@@ -24,7 +24,7 @@ namespace Ark::internal
     class SymbolExecutor : public MacroExecutor
     {
     public:
-        SymbolExecutor(MacroProcessor* macroprocessor, unsigned debug = 0);
+        using MacroExecutor::MacroExecutor;
 
         bool applyMacro(Node& node) override;
         bool canHandle(Node& node) override;
