+++ include/Ark/Compiler/Macros/Pipeline.hpp
@@ -9,8 +9,8 @@
  * 
  */
 
-#ifndef MACROS_EXECUTORS_PIPELINE_HPP
-#define MACROS_EXECUTORS_PIPELINE_HPP
+#ifndef COMPILER_MACROS_PIPELINE_HPP
+#define COMPILER_MACROS_PIPELINE_HPP
 
 #include <vector>
